+++ dflib/src/main/java/com/nhl/dflib/IntSeries.java
@@ -45,4 +45,11 @@ public interface IntSeries extends Series<Integer> {
      * null values.
      */
     IntSeries indexInt(IntPredicate predicate);
+
+
+    /**
+     * @return a IntSeries that contains non-repeating values from this Series.
+     * @since 0.6
+     */
+    IntSeries uniqueInt();
 }