+++ dflib/src/test/java/com/nhl/dflib/unit/IntSeriesAsserts.java
@@ -16,6 +16,8 @@ public class IntSeriesAsserts {
 
     public IntSeriesAsserts expectData(int... expectedValues) {
 
+        assertEquals("Unexpected Series length", expectedValues.length, data.length);
+
         for (int i = 0; i < expectedValues.length; i++) {
 
             int a = data[i];