+++ dflib/src/test/java/com/nhl/dflib/unit/LongSeriesAsserts.java
@@ -16,6 +16,8 @@ public class LongSeriesAsserts {
 
     public LongSeriesAsserts expectData(long... expectedValues) {
 
+        assertEquals("Unexpected Series length", expectedValues.length, data.length);
+
         for (int i = 0; i < expectedValues.length; i++) {
 
             long a = data[i];